requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
xmltodict==0.12.0
